Mostly font size fixing
Fixed the font size on some slides and adapted some text so it could all
be seen on the screen.
Also changed some random colors for some less random colors. And most
importantly fixed a few things that were preventing this from compiling
with an updated raspberrypilib thing.

As a bonus there's a new python based makefile (still sort of buggy),
just because makefiles are ugly.